20172096_phase1/requirements.txt
PyStemmer==1.3.0
lxml==4.2.4
